customised dashboard colours
